Attribute eager_cleanup_temp allows temp files to be purged sooner
DBICSL::Base creates a temporary directory during its constructor,
even for the times when it doesn't need one.  It uses the File::Temp
interface that adds the path to a global cleanup list that doesn't
go away until program exit.

With this change, if the user specifies eager_cleanup_temp the temp
dir only gets created if it is needed, and goes away as soon as the
SchemaLoader instance is freed.